comment

+++ src/main/java/com/forum/forum/controller/PublishController.java
@@ -30,7 +30,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String editPublish(@PathVariable(name = "id") Integer id,
+    public String editPublish(@PathVariable(name = "id") Long id,
                               Model model) {
         QuestionDTO questionDTO = questionService.findById(id);
         model.addAttribute("title", questionDTO.getTitle());
@@ -45,7 +45,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title", title);
